fix: Resolve circular import by moving _user_agent.py
