+++ sdk/ai/azure-ai-resources/azure/ai/resources/_user_agent.py
